# docs/swagger.yaml
basePath: /api
definitions:
  request.CreateTagsRequest:
    properties:
      name:
        maxLength: 10
        minLength: 1
        type: string
    required:
    - name
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
info:
  contact: {}
  description: A Tag service API in Go using GIn Framework
  title: TagService APi
  version: "1.0"
paths:
  /tags:
    get:
      description: Return list tags
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get All tags
      tags:
      - tags
    post:
      description: Save tags data in Db.
      parameters:
      - description: Create tags
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/request.CreateTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create tags
      tags:
      - tags
    put:
      description: UpdateTags data.
      parameters:
      - description: update tags by id
        in: path
        name: tagId
        required: true
        type: string
      - description: Update tags
        in: body
        name: tags
        required: true
        schema:
          $ref: '#/definitions/request.CreateTagsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: UpdateTags
      tags:
      - tags
  /tags/{tagID}:
    delete:
      description: RemoveTags By Id from data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: DeleteTags
      tags:
      - tags
  /tags/{tagId}:
    get:
      description: Return tags by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Single tags By Id
      tags:
      - tags
swagger: "2.0"